app/requirements.txt
Django==3.1.7
psycopg2-binary==2.8.5
gunicorn==20.0.4
Pillow==7.0
Requests==2.25.1
django-crispy-forms
django-comments-dab